+++ backend/webshell/consumers/__init__.py
@@ -1,0 +1,2 @@
+from .deployment_terminal import *
+from .server_terminal import *
